Move inference/type errors to utils.errors
