+++ myia/abstract/__init__.py
@@ -67,10 +67,6 @@ from .data import (  # noqa
     TYPE,
     SHAPE,
     DATA,
-    infer_trace,
-    InferenceError,
-    MyiaTypeError,
-    MyiaShapeError,
     format_abstract,
     pretty_struct,
     pretty_type,
