+++ myia/abstract/data.py
@@ -6,10 +6,8 @@ import prettyprinter as pp
 from prettyprinter.prettyprinter import pretty_python_value
 from typing import Tuple, List
 from dataclasses import dataclass
-from contextvars import ContextVar
 
-from ..debug.label import label
-from ..utils import Named, Partializable, Interned, Atom, AttrEK, \
+from ..utils import Named, Interned, Atom, AttrEK, MyiaTypeError, \
     PossiblyRecursive, OrderedSet, dataclass_methods, Cons, Empty
 from ..ir import Graph, MetaGraph
 from ..prim import Primitive
@@ -662,82 +660,6 @@ VALUE = _ValueTrack('VALUE')
 TYPE = _TypeTrack('TYPE')
 SHAPE = _ShapeTrack('SHAPE')
 DATA = _ValueTrack('DATA')
-
-
-##########
-# Errors #
-##########
-
-
-infer_trace = ContextVar('infer_trace')
-infer_trace.set({})
-
-
-class InferenceError(Exception, Partializable):
-    """Inference error in a Myia program.
-
-    Attributes:
-        message: The error message.
-        refs: A list of references which are involved in the error,
-            e.g. because they have the wrong type or don't match
-            each other.
-        traceback_refs: A map from a context to the first reference in
-            that context that fails to resolve because of this error.
-            This represents a traceback of sorts.
-
-    """
-
-    def __init__(self, message, refs=[]):
-        """Initialize an InferenceError."""
-        super().__init__(message, refs)
-        self.message = message
-        self.refs = refs
-        self.traceback_refs = infer_trace.get()
-
-
-class MyiaTypeError(InferenceError):
-    """Type error in a Myia program."""
-
-
-class MyiaShapeError(InferenceError):
-    """Shape error in a Myia program."""
-
-
-def type_error_nargs(ident, expected, got):
-    """Return a MyiaTypeError for number of arguments mismatch."""
-    return MyiaTypeError(
-        f"Wrong number of arguments for '{label(ident)}':"
-        f" expected {expected}, got {got}."
-    )
-
-
-def check_nargs(ident, expected, args):
-    """Return a MyiaTypeError for number of arguments mismatch."""
-    got = len(args)
-    if expected is not None and got != expected:
-        raise type_error_nargs(ident, expected, got)
-
-
-class TypeDispatchError(MyiaTypeError):
-    """Represents an error in type dispatch for a MetaGraph."""
-
-    def __init__(self, metagraph, types, refs=[]):
-        """Initialize a TypeDispatchError."""
-        message = f'`{metagraph}` is not defined for argument types {types}'
-        super().__init__(message, refs=refs)
-        self.metagraph = metagraph
-        self.types = types
-
-
-class TypeMismatchError(MyiaTypeError):
-    """Error to generate when expecting a type and getting another."""
-
-    def __init__(self, expected, got):
-        """Initialize a TypeMismatchError."""
-        message = f'Expected {expected}, but got {got}'
-        super().__init__(message)
-        self.expected = expected
-        self.got = got
 
 
 ##########################
