+++ myia/abstract/utils.py
@@ -9,7 +9,8 @@ from types import GeneratorType, AsyncGeneratorType
 
 from .. import dtype
 from ..utils import overload, is_dataclass_type, dataclass_methods, intern, \
-    ADT, Cons, Empty
+    ADT, Cons, Empty, MyiaTypeError, TypeMismatchError
+
 
 from .loop import Pending, is_simple, PendingTentative, \
     find_coherent_result_sync
@@ -41,8 +42,6 @@ from .data import (
     VALUE,
     TYPE,
     SHAPE,
-    MyiaTypeError,
-    TypeMismatchError,
 )
