+++ myia/ir/metagraph.py
@@ -2,10 +2,18 @@
 
 
 from .. import abstract, parser
+from ..utils import MyiaTypeError
 
 
-class GraphGenerationError(Exception):
-    """Raised when a graph could not be generated by a MetaGraph."""
+class TypeDispatchError(MyiaTypeError):
+    """Represents an error in type dispatch for a MetaGraph."""
+
+    def __init__(self, metagraph, types, refs=[]):
+        """Initialize a TypeDispatchError."""
+        message = f'`{metagraph}` is not defined for argument types {types}'
+        super().__init__(message, refs=refs)
+        self.metagraph = metagraph
+        self.types = types
 
 
 class MetaGraph:
@@ -58,8 +66,7 @@ class MultitypeGraph(MetaGraph):
 
     def normalize_args_sync(self, args):
         """Return broadened arguments."""
-        from ..abstract import broaden
-        return tuple(broaden(a) for a in args)
+        return tuple(abstract.broaden(a) for a in args)
 
     def register(self, *types):
         """Register a function for the given type signature."""
@@ -74,7 +81,7 @@ class MultitypeGraph(MetaGraph):
             if abstract.typecheck(sig, types):
                 return fn
         else:
-            raise GraphGenerationError(types)
+            raise TypeDispatchError(self, types)
 
     def generate_graph(self, args):
         """Generate a Graph for the given abstract arguments."""
@@ -82,7 +89,6 @@ class MultitypeGraph(MetaGraph):
 
     def __call__(self, *args):
         """Call like a normal function."""
-        from ..abstract import to_abstract
-        types = tuple(to_abstract(arg) for arg in args)
+        types = tuple(abstract.to_abstract(arg) for arg in args)
         fn = self._getfn(types)
         return fn(*args)
