+++ tests/test_abstract.py
@@ -7,17 +7,15 @@ import numpy as np
 from myia import dtype as ty
 from myia.prim import ops as P
 from myia.abstract import (
-    ANYTHING, MyiaTypeError,
+    ANYTHING, VALUE, TYPE, DEAD, find_coherent_result_sync,
     AbstractScalar, AbstractTuple as T, AbstractClass as AC,
     AbstractJTagged, AbstractError, AbstractFunction, AbstractUnion,
     AbstractTaggedUnion, InferenceLoop, to_abstract, build_value, amerge,
     AbstractClass, Possibilities, PendingFromList, TaggedPossibilities,
-    VALUE, TYPE, DEAD, find_coherent_result_sync,
-    abstract_clone, broaden,
-    Pending, type_to_abstract,
-    InferenceError, empty, listof
+    abstract_clone, broaden, Pending, type_to_abstract, empty, listof
 )
-from myia.utils import SymbolicKeyInstance, Cons, Empty
+from myia.utils import SymbolicKeyInstance, Cons, Empty, InferenceError, \
+    MyiaTypeError
 from myia.ir import Constant
 
 from .common import Point, to_abstract_test, i16, f32, Ty, af32_of, S, U
