+++ tests/test_api.py
@@ -5,14 +5,13 @@ import pytest
 from myia.api import myia, to_device
 from myia.cconv import closure_convert
 from myia.dtype import Bool, EnvType
-from myia.abstract import InferenceError
+from myia.abstract import ArrayWrapper
 from myia.ir import clone
 from myia.pipeline import \
     scalar_parse as parse, scalar_debug_compile as compile
 from myia.pipeline.steps import convert_arg, convert_result, NumpyChecker
 from myia.prim.py_implementations import tuple_getitem
-from myia.utils import newenv, TaggedValue
-from myia.abstract import ArrayWrapper
+from myia.utils import newenv, TaggedValue, InferenceError
 from myia.compile import load_backend, LoadingError
 
 from .common import Point, Point3D, i64, f64, to_abstract_test, ai64_of, \
