+++ tests/test_model.py
@@ -4,10 +4,11 @@ import typing
 from dataclasses import dataclass
 from numpy import ones as _ones, zeros as _zeros
 from myia.dtype import Array
-from myia.abstract import InferenceError, from_value
+from myia.abstract import from_value
 from myia.composite import grad
 from myia.pipeline import standard_pipeline
 from myia.prim.py_implementations import array_reduce, scalar_add
+from myia.utils import InferenceError
 
 from .test_compile import parse_compare
 from .test_grad import grad_test
